pyproject.toml
[tool.poetry]
name = "vendor-management-system"
version = "0.1.0"
description = ""
authors = ["sohamjani <sohamjani007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
